apps: Use actual icon name from app-info
Appstream is a gift that keeps on giving. appstream-compose sets the
icon name based on the value of <launchable> tag and not the name or ID
used by the referenced desktop file. The assumption that appid == icon
filename is wrong, even if happens to be true most of the time.

See https://github.com/flathub/flathub/pull/1949 for details.